admin put product method

+++ backend/main.js
@@ -2,6 +2,7 @@ import express from "express";
 import dotenv from "dotenv";
 import connectDb from "./config/db.js";
 import Product from "./module/product.model.js";
+import mongoose from "mongoose";
 
 //get .env to have access to the database URI
 dotenv.config();
@@ -80,17 +81,38 @@ app.post("/admin/newproduct", async (req, res)=>{
 
 app.delete("/admin/:id", async (req, res)=>{
     const {id} = req.params;
+
+    if (!mongoose.Types.ObjectId.isValid(id)){
+        return res.status(404).json({success: false, message: "Product not found"});
+    }
+
     try{
         await Product.findByIdAndDelete(id);
         return res.status(200).json({success: true, message: "Product deleted"});
     }catch(error){
         console.error(`Error deleting product: ${error.message}`);
-        return res.status(404).json({success: false, message: "Product not found"});
+        return res.status(500).json({success: false, message: "Server error"});
     }
 })
 
 app.put("/admin/:id", async (req, res)=>{
     const {id} = req.params;
+    //to get whatever admin wants to update
+    const product = req.body;
+
+    //to catch 404 case 
+    if (!mongoose.Types.ObjectId.isValid(id)){
+        return res.status(404).json({success: false, message: "Product not found"});
+    }
+
+    try{
+        //new: true so that it returns the updated product 
+        const updatedProduct = await Product.findOneAndUpdate(id, product, {new: true});
+        return res.status(200).json({success: true, data: updatedProduct});
+    }catch(error){
+        console.error(`Error updating product: ${error.message}`);
+        return res.status(500).json({success: false, message: "Server error"});
+    }
 
 })
